Changed PlanModel.productId to Int

+++ libs/fluxc/src/main/java/org/wordpress/android/fluxc/model/PlanModel.kt
@@ -1,7 +1,7 @@
 package org.wordpress.android.fluxc.model
 
 class PlanModel (
-    val productId: Long,
+    val productId: Int,
     val productSlug: String?,
     val productName: String?,
     val isCurrentPlan: Boolean,
